Add outer constructors for "state" types
When playing around at the REPL, it's nice to not need to specify
the type-parameter in state objects. This provides an outer constructor
that fills the parameters in from the arguments.

One missing case is SmartsParserState, but that seems to be a good case
for forcing the user to supply them directly.